Add files via upload
This project is meant to assess the level of progress made at the state level as well in the metropolitan cities, in terms of containing the notorious and wildfire spread of the Covid-19 Virus across the land of America. Though we are in the earliest phase of the post-covid  era, this  model is highly recommended for the stakeholders in  the healthcare sector in containing any possible outbreak of a novel disease,   potentially able to take the dimension and magnitude of the notorious Covid-19 Virus of which the world is still learning about till today.